requirements.txt
agentops>=0.3.0
requests>=2.28.0
biopython>=1.79
beautifulsoup4>=4.11.0
numpy>=1.21.0
httpx>=0.23.0
# Note: PyMOL should be installed separately:
# macOS: brew install brewsci/bio/pymol
# Linux: sudo apt-get install pymol
# Windows: Download from https://pymol.org/2/ 